ERC4626 vault that invests in eTokens
ETokensBundleVault is an ERC-4626 vault that invests the funds received
in several underlying eTokens (all belonging to the same pool).

The money inflows (deposits) are distributed among the eTokens in configurable percentages.

The withdrawals are distributed among the eTokens based on the percentage of the
total value of the vault.

WIP: not tested yet.